'Refactored by Sourcery'

+++ contrib/opencensus-ext-azure/opencensus/ext/azure/common/storage.py
@@ -43,7 +43,7 @@ class LocalFileBlob(object):
 
     def put(self, data, lease_period=0):
         try:
-            fullpath = self.fullpath + '.tmp'
+            fullpath = f'{self.fullpath}.tmp'
             with open(fullpath, 'w') as file:
                 for item in data:
                     file.write(json.dumps(item))
@@ -53,7 +53,7 @@ class LocalFileBlob(object):
                     file.write('\n')
             if lease_period:
                 timestamp = _now() + _seconds(lease_period)
-                self.fullpath += '@{}.lock'.format(_fmt(timestamp))
+                self.fullpath += f'@{_fmt(timestamp)}.lock'
             os.rename(fullpath, self.fullpath)
             return self
         except Exception:
@@ -64,7 +64,7 @@ class LocalFileBlob(object):
         fullpath = self.fullpath
         if fullpath.endswith('.lock'):
             fullpath = fullpath[: fullpath.rindex('@')]
-        fullpath += '@{}.lock'.format(_fmt(timestamp))
+        fullpath += f'@{_fmt(timestamp)}.lock'
         try:
             os.rename(self.fullpath, fullpath)
         except Exception:
@@ -93,8 +93,9 @@ class LocalFileStorage(object):
         self._maintenance_task = PeriodicTask(
             interval=self.maintenance_period,
             function=self._maintenance_routine,
-            name='{} Storage Worker'.format(source)
+            name=f'{source} Storage Worker',
         )
+
         self._maintenance_task.daemon = True
         self._maintenance_task.start()
 
@@ -116,7 +117,7 @@ class LocalFileStorage(object):
             # Race case will throw OSError which we can ignore
             pass
         try:
-            for blob in self.gets():
+            for _ in self.gets():
                 pass
         except Exception:
             pass  # keep silent
@@ -130,14 +131,13 @@ class LocalFileStorage(object):
             path = os.path.join(self.path, name)
             if not os.path.isfile(path):
                 continue  # skip if not a file
-            if path.endswith('.tmp'):
-                if name < timeout_deadline:
-                    try:
-                        os.remove(path)
-                        logger.warning(
-                            'File write exceeded timeout. Dropping telemetry')
-                    except Exception:
-                        pass  # keep silent
+            if path.endswith('.tmp') and name < timeout_deadline:
+                try:
+                    os.remove(path)
+                    logger.warning(
+                        'File write exceeded timeout. Dropping telemetry')
+                except Exception:
+                    pass  # keep silent
             if path.endswith('.lock'):
                 if path[path.rindex('@') + 1: -5] > lease_deadline:
                     continue  # under lease
